Tweaked a few stuff, for example added a back button everywhere for these features and extracted common components into shared folder
